feat: refactor RoomStageStateMachine to handle stage transitions and initialization callbacks
