+++ src/main/java/com/icebreaker/be/application/room/statemachine/RoomStageTransitionKey.java
@@ -1,7 +1,7 @@
 package com.icebreaker.be.application.room.statemachine;
 
+import com.icebreaker.be.application.room.event.StageTransitionEvent;
 import com.icebreaker.be.domain.room.vo.Stage;
-import com.icebreaker.be.domain.room.vo.StageTransitionEvent;
 
 public record RoomStageTransitionKey(Stage stage, StageTransitionEvent event) {
